Move PromseQueue into utils.
